ut_fix

+++ test/ut/testcase/ut_ffrt_io.cpp
@@ -278,7 +278,7 @@ HWTEST_F(ffrtIoTest, ffrt_timer_start_succ_short_timeout_flagwake, TestSize.Leve
         }, {}, {});
     ffrt_epoll_ctl(qos, EPOLL_CTL_DEL, testFd, 0, nullptr, nullptr);
     ffrt::wait();
-    EXPECT_EQ(1, x);
+    EXPECT_EQ(static_cast<int>(1), x);
     close(testFd);
 }
